Матеріали/33_Flask_blog/pyproject.toml
[tool.poetry]
name = "33-flask-blog"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
readme = "README.md"
packages = [{include = "33_flask_blog"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
